+++ src/main/java/com/scanoss/cli/ScanCommandLine.java
@@ -39,7 +39,7 @@ import static com.scanoss.cli.CommandLine.printMsg;
  * Provide a Scan CLI subcommand to interacting with the SCANOSS Java SDK
  * </p>
  */
-@SuppressWarnings("unused")
+@SuppressWarnings({"unused", "CanBeFinal"})
 @picocli.CommandLine.Command(name = "scan", description = "Scan the given file/folder/wfp")
 class ScanCommandLine implements Runnable {
     @picocli.CommandLine.Spec
@@ -47,8 +47,41 @@ class ScanCommandLine implements Runnable {
 
     @picocli.CommandLine.Option(names = {"-h", "--help"}, usageHelp = true, description = "Display help information")
     private boolean helpRequested = false;
+
+    @picocli.CommandLine.Option(names = {"-S", "--skip-snippets"}, description = "Skip the generation of snippets")
+    private boolean skipSnippets = false;
+
+    @picocli.CommandLine.Option(names = "--all-extensions", description = "Scan all file extensions")
+    private boolean allExtensions = false;
+
+    @picocli.CommandLine.Option(names = "--all-hidden", description = "Scan all hidden files/folders")
+    private boolean allHidden = false;
+
+    @picocli.CommandLine.Option(names = "--all-folders", description = "Scan all folders")
+    private boolean allFolders = false;
+
+    @picocli.CommandLine.Option(names = {"-T", "--threads"}, description = "Number of parallel threads to use")
+    private int numThreads = 5;
+
+    @picocli.CommandLine.Option(names = "--apiurl", description = "SCANOSS API URL (optional - default: https://osskb.org/api/scan/direct)")
+    private String apiUrl;
+
+    @picocli.CommandLine.Option(names = {"-k", "--key"}, description = "SCANOSS API Key token (optional - not required for default OSSKB URL)")
+    private String apiKey;
+
+    @picocli.CommandLine.Option(names = {"-R", "--retry"}, description = "Retry limit for API communication (optional - default 5)")
+    private int retryLimit = 5;
+
+    @picocli.CommandLine.Option(names = {"-M", "--timeout"}, description = "Timeout (in seconds) for API communication (optional - default 120)")
+    private int timeoutLimit = 5;
+
+    @picocli.CommandLine.Option(names = {"-F", "--flags"}, description = "Scanning engine flags (1: disable snippet matching, 2 enable snippet ids, 4: disable dependencies, 8: disable licenses, 16: disable copyrights, 32: disable vulnerabilities, 64: disable quality, 128: disable cryptography,256: disable best match only, 512: hide identified files, 1024: enable download_url, 2048: enable GitHub full path, 4096: disable extended server stats")
+    private String scanFlags;
+
     @picocli.CommandLine.Parameters(arity = "1", description = "file/folder to scan")
     private String fileFolder;
+
+    private Scanner scanner;
 
     /**
      * Run the 'scan' command
@@ -56,12 +89,33 @@ class ScanCommandLine implements Runnable {
     @Override
     public void run() {
         var err = spec.commandLine().getErr();
-        if (CommandLine.debug) {
-            printDebug(err, "Debug enabled.");
-        }
         if (fileFolder == null || fileFolder.isEmpty()) {
             throw new RuntimeException("Error: No file or folder specified to scan");
         }
+        if (com.scanoss.cli.CommandLine.debug) {
+            if (numThreads != 5) {
+                printMsg(err, String.format("Running with %d threads.", numThreads));
+            }
+            if (timeoutLimit != 120) {
+                printMsg(err, String.format("Scanning with timeout of %d seconds", timeoutLimit));
+            }
+            if (skipSnippets) {
+                printDebug(err, "Skipping snippets.");
+            }
+            if (allExtensions) {
+                printMsg(err, "Scanning all file extensions/types.");
+            }
+            if (allHidden) {
+                printDebug(err, "Scanning all hidden files/folders.");
+            }
+            if (scanFlags != null && !scanFlags.isEmpty()) {
+                printMsg(err, String.format("Using flags %s", scanFlags));
+            }
+        }
+        scanner = Scanner.builder().skipSnippets(skipSnippets).allFolders(allFolders).allExtensions(allExtensions)
+                .hiddenFilesFolders(allHidden).numThreads(numThreads).url(apiUrl).apiKey(apiKey)
+                .retryLimit(retryLimit).timeout(timeoutLimit).scanFlags(scanFlags)
+                .build();
         File f = new File(fileFolder);
         if (!f.exists()) {
             throw new RuntimeException(String.format("Error: File or folder does not exist: %s\n", fileFolder));
@@ -83,7 +137,6 @@ class ScanCommandLine implements Runnable {
     private void scanFile(String file) {
         var out = spec.commandLine().getOut();
         var err = spec.commandLine().getErr();
-        Scanner scanner = Scanner.builder().build();
         try {
             printMsg(err, String.format("Scanning %s...", file));
             String result = scanner.scanFile(file);
@@ -98,6 +151,8 @@ class ScanCommandLine implements Runnable {
                 e.printStackTrace(err);
             }
             throw e;
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
         }
         throw new RuntimeException(String.format("Something went wrong while scanning %s", file));
     }
@@ -110,7 +165,6 @@ class ScanCommandLine implements Runnable {
     private void scanFolder(String folder) {
         var out = spec.commandLine().getOut();
         var err = spec.commandLine().getErr();
-        Scanner scanner = Scanner.builder().build();
         try {
             printMsg(err, String.format("Scanning %s...", folder));
             List<String> results = scanner.scanFolder(folder);