+++ CHANGELOG.md
@@ -6,12 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
+
 - Upcoming changes...
 
+## [0.2.0] - 2023-07-04
 
-## [0.0.1] - 2023-06-??
 ### Added
-- ?
 
-[0.0.1]: https://github.com/scanoss/scanoss.java/compare/v0.0.0...v0.0.1
+- First pass at the following Classes
+    - Fingerprinting ([Winnowing](src/main/java/com/scanoss/Winnowing.java))
+    - Scanning ([Scanner](src/main/java/com/scanoss/Scanner.java))
+    - REST Interface ([ScanApi](src/main/java/com/scanoss/rest/ScanApi.java))
+    - JSON Utils ([JsonUtils](src/main/java/com/scanoss/utils/JsonUtils.java))
+- CLI ([CommandLine](src/main/java/com/scanoss/cli/CommandLine.java))
+
+[0.0.1]: https://github.com/scanoss/scanoss.java/compare/v0.0.0...v0.2.0